Move DetailsCard to components folder as it is used several places
