+++ src/components/CommPkgDetailsCard/DetailsCard.test.tsx
