+++ src/components/CommPkgDetailsCard/DetailsCard.tsx
